+++ ios/RCTBackgroundGeolocation/RCTBackgroundGeolocation.h
@@ -6,7 +6,11 @@
 //  Copyright © 2016 mauron85. All rights reserved.
 //
 
+#if __has_include("RCTBridgeModule.h")
 #import "RCTBridgeModule.h"
+#else
+#import <React/RCTBridgeModule.h>
+#endif
 #import "LocationManager.h"
 
 @interface RCTBackgroundGeolocation : NSObject <RCTBridgeModule, LocationManagerDelegate>